Fix build issues in content/ for Android x64
BUG=346626

Review URL: https://codereview.chromium.org/182833003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@254136 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b24792554c8bd12709afc1134d2558b0eab79e40

+++ content/browser/android/content_view_core_impl.cc
@@ -1666,7 +1666,7 @@ bool ContentViewCoreImpl::GetUseDesktopUserAgent(
   return entry && entry->GetIsOverridingUserAgent();
 }
 
-void ContentViewCoreImpl::UpdateImeAdapter(int native_ime_adapter,
+void ContentViewCoreImpl::UpdateImeAdapter(long native_ime_adapter,
                                            int text_input_type,
                                            const std::string& text,
                                            int selection_start,
