+++ content/browser/android/content_view_core_impl.h
@@ -255,7 +255,7 @@ class ContentViewCoreImpl : public ContentViewCore,
                        const gfx::Vector2dF& content_offset,
                        float overdraw_bottom_height);
 
-  void UpdateImeAdapter(int native_ime_adapter, int text_input_type,
+  void UpdateImeAdapter(long native_ime_adapter, int text_input_type,
                         const std::string& text,
                         int selection_start, int selection_end,
                         int composition_start, int composition_end,
